app/requirements.txt
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
fastapi
uvicorn
psycopg2
matplotlib